requirement.txt
jsonrpclib-pelix
lxml
newspaper3k
pandas
pika
python-dateutil
pymongo
redis
requests
sklearn
tensorflow
watchdog
yaml
